Handled path agnostic hashing with job distribution
The root path to substitute in hash inputs is serialized over the network to the remote workers.
Bumped protocol version to handle transition with previous versions which don't serialize this info.
Only ObjectNode is explicitely handling distribution by looking in the job payload for the root path instead of global state.